Consolidates stats unit tests
